Use the resource plugin to copy the compatibility files instead of adding the source directory. This seems more compatible
