+++ curator-x-async/src/test/java/org/apache/curator/x/async/CompletableBaseClassForTests.java
@@ -20,8 +20,7 @@ package org.apache.curator.x.async;
 
 import com.google.common.base.Throwables;
 import org.apache.curator.test.BaseClassForTests;
-import org.apache.curator.test.Timing;
-import org.apache.curator.test.Timing2;
+import org.apache.curator.test.compatibility.Timing2;
 import org.testng.Assert;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.ExecutionException;
